chore: rename base commands
